frameworks:av: Add secure-record feature to bypass secure encoder codec in CTS
Secure encoder tests are being executed in CTS because the entry
is included in media_codecs.xml file. As the CTS app is
providing a non-secure buffer, kernel panic/H/W freeze are being
observed randomly. In order to bypass the secure encoder tests, add
secure-record as a valid feature for encoders. This feature will
be defined in the secure encoder entry in xml file and will be
ignored when creating the codec for being an advanced feature

Change-Id: I50fd2641e389572bdca66bc78961492a1e65fcb7