frameworks:av: Use different XMLs for pure & VA
Features like secure encoder which require
changes in AOSP framework will only be enabled in
value add builds hence, separate the XMLs for pure
& value add builds to pass the CTS/VTS
Fallback to the default XML if *_va.xml is not
present.

The default XML names will be overwritten with *_va.xml
in value add builds. This change will only be present in
VA non-GSI builds.

Change-Id: I0f3af676d72293488e420d562c801c8af2605747