Cache MMAP client silenced state.
This changes Sept ASB gerrit reference https://review-android.quicinc.com/#/c/4654995/

When starting MMAP input stream, APM will check if the client is allowed
to capture at that moment or not and call setRecordSilenced if the
client is not allowed. However, the client is not active when starting
the MMAP input stream. In that case, the client silenced state will be
lost and the client will be able to capture even though it is not
allowed. In this CL, when setRecordSilenced is called, it will cache
the client silenced state so that it can apply when the client is
active.

Test: atest AAudioTests
Test: repo steps from the bug
Bug: 235850634
Merged-In: I49b5a0f08d1747053f868db6e88c0f677256fc3c
(cherry picked from commit 0960903b2fee5d1d449ffcd598e0b5d3a945d99a)
(cherry picked from commit cfc10a47f66b2864350aa31a9bf96b76f35b366b)

Change-Id: Ib39c4a6620143851031300e7b554bf1bbd9d6e79