Media: libstagefright: notify the exceeds message in advance
in this project, mpeg4 node and the video record sound play
consume long duration. it caused the EOS cannot be transfer
to camera frameowork on time and then cannot be transfer
to camera hal. finally this cause the camera EIS node buffer
cannot be return back to framework and then stop video recording
failed. this change used to send notify message in advance to
make app get that message earlier.

Change-Id: If4a87cbafab04283629ea58a58e2f054368624b5