StagefrightRecorder: Reset AudioSourceNode during LPCM recording.
AudioSource is created as part of ExtendedAudioSource for LPCM
recording. AudioSourceNode is not set to the right AudioSource
in this case. Amplitude is not shown in SoundRecorder application
due to this reason
Fix is to reset AudioSourceNode to the instance that is returned
by setPCMRecording
CRs-Fixed: 2068175
Change-Id: I13e6a78b8bdb1e5a23c398508100617e6fc8201a

StagefrightRecorder: add null check for recording source
Check that the recording source returned is not null before
trying to add it to the writer. Writer will not check
that this source is not null before dereferencing it.
CRs-Fixed: 1085154
Change-Id: I8dda57f39bbf5e289facc5ce261d53d988630620

libmedia: correct latency computing of AudioRecord
- Originally, audio record latency is calculated by frame
  buffers allocated for this track, while the actual
  latency is determined by audio hal.
- Compute the track latency by frame count returned
  from audio hal when in TRANSFER_CALLBACK mode
CRs-fixed: 1062825
Change-Id: I26e5e47e8cc3720895b962f7aab8a595a54b7c83

libmedia : Fix for AudioRecord CTS test case
  -if sample rate is not specificed during construction of audio record
   it updates sampling rate with default value based on the source.
   but divide by zero execption is happeing since
   it uses same unspeficied sampling rate to calculate latency.
  -Use updated default sample rate to calcuate latency vaulue.
CRs-Fixed: 1090804
Change-Id: Icb1e717d869de7bcd4b52194f3ee4e8b0b318da4

StagefrightRecorder: Reset AudioSourceNode during LPCM recording.
AudioSource is created as part of ExtendedAudioSource for LPCM
recording. AudioSourceNode is not set to the right AudioSource
in this case. Amplitude is not shown in SoundRecorder application
due to this reason
Fix is to reset AudioSourceNode to the instance that is returned
by setPCMRecording
CRs-Fixed: 2068175
Change-Id: I13e6a78b8bdb1e5a23c398508100617e6fc8201a

audio: calculate the max buffer size runtime
Currently kMaxBufferSize is predefined as 30720,
but this is valid only for camcorder use-case and
not for other use-case, but as this is predefined
calculations for other than camera use-case go
wrong, such as frameCount and record latency, which
leads to issues such as data drop at the begining
due to high record latency.
So instead predefining the MaxBufferSize calculate
it only for camcorder use-case and use standard AOSP
value otherwise.
Change-Id: Ie7d43ddfdcb2b13a8a5eaa12cc8a233ac0d9470d

Change-Id: I9f489564e35918d91e3034fc565320d0cf2d97be