Nuplayer: Handle get input/output format errors cleanly
In NuplayerDecoder, after MediaCodec::configure() returns success,
it is assumed that MediaCodec is in CONFIGURED state.
But, after MediaCodec::configure() has returned success, there are
some race conditions in CCodecWatchdog thread that can move
MediaCodec to error state after configuration.
So, errors from MediaCodec::getInputFormat() and getOutputFormat()
need to be handled cleanly instead of assert checks.
This removal of assert check is temporary fix to avoid tombstones
till the race condition is fixed in CCodec.

Change-Id: I4d9e446b17c04e1abf87899eb89f161597c4da03