Nuplayer: Use video render rate from video decoder
If the video decoder has updated the output render rate,
use that value for scheduling instead of the value which
was configured to decoder.

Decoder might not use the configured render rate value
as is, if the configured value is more than the supported
value. In such scenarios, if the decoder has updated the
output render rate, use it for video frame scheduling.

Change-Id: Ia21f53ff5226ef5a4cec2209be1f32bff89ef1a1
(cherry picked from commit 077099c64472577cf7b42c76f4c05062a0cdc7b9)