Camera2 multi-client: Store clients based on PID.
In camera2 multi-client, currently clients are stored in the AidlCamera3SharedDevice
based on uid. That should be changed to PID as the same uid can have multiple processes
which can access the same camera.


Bug:389159564
Test: run CTS test SharedCameraTest#testCameraDeviceSharingSupported and
verify in the logcat that HAL level camera disconnect only happens when
last client of the camera disconnects.

Change-Id: I6300426caf5bc0a1033e983d0ff3c552e45b625b