Camera multi-client fix bug in notifyError
In camera multi-client feature a check was added to deliver notifications to the clients only if the reqid is one of the
requests sent by them. For some of the notifications like disconnect in
notifyError, the error does not correspond to any specific request id. Hence the reqid matching check should be skipped for such errors when req id is -1.


Bug: 392222024
Test: run CTS test SharedCameraTest#testOpenMixedMode
Change-Id: I13fb01789f3e8d86369b4366f92990cb1eeed633