C2SoftMpeg4Dec: Set RGBX output format.
The default output format for the ADP platform is RGB, which is the only format that can be displayed due to virtio GPU limitations.
The default output format for Google MPEG4 software codecs is multi-planar YUV420, which is then converted to YV12 (the final output of the software codec).
To meet virtio GPU requirements, switch the MPEG4 decoder output to RGB and provide the necessary format conversion.
The libyuv library has been added as a format converter.
Extended the list of supported pixel formats with RGBX_8888 - the default for the ADP platform.
A getter for the output pixel format was added. Use the current value to fetch the correct output graphics buffer.
Before finishing frame decoding, convert the YUV420 output buffer to the desired output format.

Issue: E0997-2576
Change-Id: Ie4688cca34435ad3a73780b169f5d22a341b520d
Signed-off-by: Artur Tynecki <Artur.Tynecki.external@opensynergy.com>
(cherry picked from commit aecb0c60d367d0689c20576a9da62351117c7f87)